Night wrapup with TODOs
So far, I am able to use Route to switch pages with the URL updated, Load content from the BooksAPI and get a handle on which selector is clicked.

My TODOs are to get the state to be updated locally, save the state to the API then start working on the search component.